+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### **0.2.1**
 - Added a refresh token loop that can be triggered from `Session`
 - Added Alarms
+- Added the ability to log out
 
 ### **0.2.0**
 - First proper release